backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.8.0
websockets>=12.0
python-multipart>=0.0.6
aiohttp
strands-agents